backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Telegram API
telethon==1.34.0
cryptg==0.4.0

# Database
asyncpg>=0.29.0
supabase==2.3.4
postgrest==0.13.2

# Authentication (PyJWT replaces unmaintained python-jose)
PyJWT[crypto]>=2.8.0

# Encryption
cryptography==42.0.0

# Environment
python-dotenv==1.0.0

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.7.0

# HTTP client (used for Supabase Broadcast API)
httpx>=0.27.0

# Retry
tenacity>=8.2.0

# Structured logging
python-json-logger>=2.0.0

# Error tracking
sentry-sdk>=2.0.0

# Email
resend==0.7.0

# Date/Time
python-dateutil==2.8.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
